refactor(app): extract useExtensionsHydration hook
Phase 1 step 7 of the god-file cleanup plan. Move the bridge-snapshot
hydration surface out of App.tsx into
src/hooks/useExtensionsHydration.ts. The hook owns all the wholesale-
replacement state for extension-contributed surfaces:

- themesRef / layoutCatalogueRef / extensionEventRoutesRef /
  extensionEventRoutingModeRef / extensionKeybindingsRef /
  frontendModulesRef / slashCommandsRef / extensionSlashNamesRef.
- lastExtensionStateKeysRef (JSON-pointer pruning ledger).
- ExtensionTheme type + BUILTIN_THEMES const.
- injectThemeStyle, hydrateThemes, hydrateExtensions,
  hydrateEventRoutes, hydrateKeybindings, hydrateExtensionLayouts,
  hydrateFrontendModules, hydrateSlashCommands.
- listThemes, summarizeLayoutComponents, activateLayoutById.
- The mount-only seed effect for /slashCommands + /sidebar/layouts.
- The /sidebar/components effect that re-walks the layout on change.

App passes setState/setLayout/stateRef/registry/appendSystem/layout
as ctx. The hook's hydrate* surfaces match the existing call sites
in the bridge message handler 1:1 — no rename needed.

App.tsx 4978 -> 4588 lines (-390). Total: 6224 -> 4588 (-1636,
-26.3%). tsc clean, eslint clean, 305/305 tests pass. No behavior
change.
